requirements.txt
# requirements.txt
pdfminer.six==20221105
